requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0 